==> R/Data Analysis/Unsupervised Learning/2b_Perfect_Order_ICA.Rmd <==
---
title: "2b Independent Components Analysis of Perfect Order Data"
date: '`r format(Sys.Date(), "%B %d, %Y")`'
output: pdf_document
---

To Do
-----
* read fastICA documentation
* Do the analysis for a range of n.comps to see how the results differ
* Determine how to interpret the results
* What's the best k?

Observations
------------
* 

```{r packages, warning=FALSE}
# Load necessary packages 

library(fastICA)
library(lattice)
library(ggplot2)
library(moments)
library(microbenchmark)


```

```{r ImportData, echo=FALSE}
# Clear workspace and set working directory
rm(list = ls())
setwd("/Users/kbrooks/Documents/CS7641 Machine Learning/Assignment 3")
par(mar = c(0, 0, 0, 0))

d = readRDS("models/Perfect_Order_cluster.rds")

set.seed(250)
#summary(d)

```


```{r dataAnalysis, echo=FALSE}

# verify by plotting mean of columns

mar <- par()$mar
par(mar=mar+c(0,5,0,0))
barplot(sapply(d, mean), horiz=T, las=1, cex.names=0.8, main = "Mean")
par(mar=mar)

dev.copy(png,'images/Perfect_Order_ica_mean.png')
dev.off()

# verify by plotting variance of columns

mar <- par()$mar
par(mar=mar+c(0,5,0,0))
barplot(sapply(d, var), horiz=T, las=1, cex.names=0.8, main = "Variance")
par(mar=mar)

dev.copy(png,'images/Perfect_Order_ica_variance.png')
dev.off()

# verify by plotting skewness of columns

mar <- par()$mar
par(mar=mar+c(0,5,0,0))
barplot(sapply(d, skewness), horiz=T, las=1, cex.names=0.8, main = "Skewness")
par(mar=mar)

dev.copy(png,'images/Perfect_Order_ica_skewness.png')
dev.off()

# verify by plotting kurtosis of columns

mar <- par()$mar
par(mar=mar+c(0,5,0,0))
barplot(sapply(d, kurtosis), horiz=T, las=1, cex.names=0.5, main = "Kurtosis")
par(mar=mar)

dev.copy(png,'images/Perfect_Order_ica_kurtosis.png')
dev.off()


```




Independent Components Analysis
-------------------------------

This analysis implements ICA using the _fastICA_ package in R[^fn-1]. _fastICA_ is tuned using the following parameters:

* X - a data matrix with n rows representing observations and p columns representing variables.
* n.comp - number of components to be extracted
* alg.typ	- if alg.typ == "parallel" the components are extracted simultaneously (the default). if alg.typ == "deflation" the components are extracted one at a time.
* fun	- the functional form of the G function used in the approximation to neg-entropy (see ‘details’).
* alpha	- constant in range [1, 2] used in approximation to neg-entropy when fun == "logcosh"
* method - if method == "R" then computations are done exclusively in R (default). The code allows the interested R user to see exactly what the algorithm does. if method == "C" then C code is used to perform most of the computations, which makes the algorithm run faster. During compilation the C code is linked to an optimized BLAS library if present, otherwise stand-alone BLAS routines are compiled.
* row.norm - a logical value indicating whether rows of the data matrix X should be standardized beforehand.
* maxit	- maximum number of iterations to perform.
* tol	- a positive scalar giving the tolerance at which the un-mixing matrix is considered to have converged.
* verbose	- a logical value indicating the level of output as the algorithm runs.
* w.init- Initial un-mixing matrix of dimension c(n.comp, n.comp). If NULL (default) then a matrix of normal r.v.'s is used.

[^fn-1]: J L Marchini, C Heaton and B D Ripley (2013). fastICA: FastICA Algorithms to perform ICA and Projection Pursuit. R package version 1.2-0. http://CRAN.R-project.org/package=fastICA


```{r runpca, warning=FALSE}

#?fastICA()

pc.cor <- princomp(d,cor=T)

ica.5 <- fastICA(d,n.comp=5)
saveRDS(ica.5, "models/Perfect_Order_ICA_5_model.rds")
ica.10 <- fastICA(d,n.comp=10)
saveRDS(ica.10, "models/Perfect_Order_ICA_10_model.rds")
ica.15 <- fastICA(d,n.comp=15)
saveRDS(ica.15, "models/Perfect_Order_ICA_15_model.rds")
ica.20 <- fastICA(d,n.comp=20)
saveRDS(ica.20, "models/Perfect_Order_ICA_20_model.rds")


par(mar = c(0, 2, 0, 0), mfrow=c(1,4))
barplot(sapply(data.frame(ica.5$S), kurtosis), horiz=T, las=1, cex.names=0.5, main = "k = 5")
barplot(sapply(data.frame(ica.10$S), kurtosis), horiz=T, las=1, cex.names=0.5, main = "k = 10")
barplot(sapply(data.frame(ica.15$S), kurtosis), horiz=T, las=1, cex.names=0.5, main = "k = 15")
barplot(sapply(data.frame(ica.20$S), kurtosis), horiz=T, las=1, cex.names=0.5, main = "k = 20")
par(mar=mar)

dev.copy(png,'images/Perfect_Order_ICA_plot.png')
dev.off()

biplot(pc.cor)

S <- ica$S
dimnames(S) <- list(dimnames(d)[[1]],paste("Component",1:22,sep=""))

A <- ica$A
dimnames(A) <- list(dimnames(d)[[2]],paste("Cmp.",1:22,sep=""))

# windows()

biplot(S[,1:2],A[,1:2])

                    


```


```{r icaPlots}
par(mfrow = c(1, 3))
plot(ica.20$X, main = "Pre-processed data")
plot(ica.20$X %*% ica.20$K, main = "PCA components")
plot(ica.20$S, main = "ICA components")
```


```{r ICATiming}

library(fastICA)
library(microbenchmark)

d = readRDS("models/Perfect_Order_cluster.rds")

trainingTime = microbenchmark(
    ica.15 <- fastICA(d,n.comp=15), 
    times = 10L)

cat("Mean training time:", trainingTime$time[3]/1000000000, "ns")
```
